Switch to one day trailing APR
Calculates the APR by using the OUSD rebase ratio.

This has the upside that it's simple to calculate and exactly matches
user's balance changes.

It has the downside that the number it pulls from only gets updated
on rebases, making this method less acurate. It's bit iffy using it
on only one day, but that's the data we have at the moment.